+++ src/main/java/com/elmakers/mine/bukkit/plugins/magic/CleanupBlocksTask.java
@@ -1,20 +1,24 @@
 package com.elmakers.mine.bukkit.plugins.magic;
 
 import com.elmakers.mine.bukkit.dao.BlockList;
+import com.elmakers.mine.bukkit.utilities.UndoQueue;
 
 public class CleanupBlocksTask implements Runnable
 {
 	protected BlockList undoBlocks;
 	protected Spells spells;
+	protected UndoQueue queue;
 
-	public CleanupBlocksTask(Spells spells, BlockList cleanup)
+	public CleanupBlocksTask(UndoQueue queue, Spells spells, BlockList cleanup)
 	{
 		this.undoBlocks = cleanup;
 		this.spells = spells;
+		this.queue = queue;
 	}
 
 	public void run()
 	{
 		this.undoBlocks.undo(spells);
+		queue.removeScheduledCleanup(undoBlocks);
 	}
 }