+++ src/main/java/com/elmakers/mine/bukkit/plugins/magic/spells/FrostSpell.java
@@ -140,7 +140,7 @@ public class FrostSpell extends Spell
 
 		BlockList frozenBlocks = action.getBlocks();
 		frozenBlocks.setTimeToLive(timeToLive);
-		spells.scheduleCleanup(frozenBlocks);
+		spells.scheduleCleanup(player.getName(), frozenBlocks);
 		castMessage("Frosted " + action.getBlocks().size() + " blocks");
 
 		return SpellResult.SUCCESS;